chore: move InfiniteHits stories to external file
